src/utils/pyproject.toml
[project]
name = "utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.31",
    "datasets>=4.1.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]
